new Gizmo.cpp file
